djangoProject/requirements.txt
Django
requests
djangorestframework
django-rest-knox